buildwise-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "buildwise-backend"
version = "0.1.0"
description = "BuildWise - Building Energy Simulation SaaS Backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",

    # IDF generation
    "eppy>=0.5.63",
    "jinja2>=3.1.4",

    # Validation
    "pydantic>=2.10.0",
    "jsonschema>=4.23.0",

    # Config
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",

    # Cloud
    "google-cloud-storage>=2.18.0",

    # Utilities
    "structlog>=24.4.0",

    # ems_simulation bridge dependencies
    "pandas>=2.2.0",

    # AI (NL → BPS parsing)
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# IDF template strings are necessarily long
"app/services/idf/hvac_vav.py" = ["E501"]
# Test utility scripts with long formatted output
"tests/verify_full_ems_data.py" = ["E501"]
"tests/verify_idf_patcher.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--ignore=tests/docker_integration_test.py --ignore=tests/docker_e2e_test.py"

[tool.mypy]
python_version = "3.12"
strict = true
